==> 2_Aggregate_Demand_Spark.ipynb <==
# 2  Calculating the aggregate demand with Spark

In this section we will explore our full dataset with a Spark mindset. We have already explored it in the introduction with pandas to know what it contains and how it looks like but now we need to prepare it to be able to do the calculations and aggregations we need with the appropriate data types. We will start with the data sample and then try the scheme with the full dataset. As a result, we will build a program that can be run in HDFS.

#### Downloading the data

#Sample
!wget 'https://files.datapress.com/london/dataset/smartmeter-energy-use-data-in-london-households/UKPN-LCL-smartmeter-sample.csv'

#Full dataset
!wget 'https://files.datapress.com/london/dataset/smartmeter-energy-use-data-in-london-households/Power-Networks-LCL-June2015(withAcornGps).zip'

#### File transformation to bz2

The file is compressed in .zip format. The problem of this format is it cannot be broken into pieces without being corrupted. Therefore, if we attempt a MapReduce operation we will get a useless binary output. To be able to use it without decompressing it (as we do not have enough disk space) we will transform it line by line to bzip2 using pipes.

!unzip -c 'Power-Networks-LCL-June2015(withAcornGps).zip'  | bzip2 > 'Power-Networks-LCL-June2015.bz2'

#### Library importing

Now we are ready to start. The commented lines would be needed if our Virtual Machine had not been already configured to make our life easier to import SparkContext and assign it to sc upon start.

#from pyspark import SparkContext
#sc = SparkContext()
from pyspark.sql.types import DateType
from datetime import datetime
import numpy as np
from pyspark import Row
sc

sc.version

#### Sample dataset

We will discover the steps needed to process our dataset using the small sample dataset.

path_sample = 'UKPN-LCL-smartmeter-sample.csv'

data_sample = sc.textFile(path_sample)

data_sample.take(3)

The header is useful in pandas but we need to remove it when working with Spark.

header = data_sample.first()

data_sample2 = data_sample.filter(lambda l: l != header).persist()

data_sample2.take(3)

Now we will map the file extracting the relevant fields and converting them to the appropriate types. We will ignore the last field, which is actually a higher-level ACORN grouping compared to the second last. ACORN is a UK consumer socioeconomical segmentation system that may be useful in this work (see http://acorn.caci.co.uk/downloads/Acorn-User-guide.pdf).

def line2tuple(l):
    fields = l.split(',')
    if len(fields) == 6:
        ID = fields[0]
        tariff = fields[1]
        DateTime = datetime.strptime(fields[2], '%d/%m/%Y %H:%M:%S')
        ACORN = fields[4]
        try:
            consumption = float(fields[3])
        except ValueError:
            consumption = np.nan
        return (ID, tariff, DateTime, consumption, ACORN)
    else:
        return (np.nan,np.nan,np.nan,np.nan,np.nan)

data_sample2.map(line2tuple).take(3)

rows_sample = data_sample2.map(line2tuple)\
.map(lambda x: Row(ID = x[0], Tariff = x[1], DateTime = x[2], kWh_30min = x[3], ACORN = x[4]))

df_sample = rows_sample.toDF()

df_sample.show(5)

We can try calculating the mean grouping by tariff. In this case, there is only one consumer, which is subject to the standard tariff.

df_sample.dropna().groupBy('Tariff').mean().collect()

#Checking the types, everything looks correct.
df_sample.printSchema

##### Saving the data to a text file

At the end of our full dataset processing, we will aggregate the data and and save it to a text file. Let us explore alternatives with the smaller sample dataset.

There are different options to save the data as a text file. The problem with these first two options is that it will save it in several text files, as it does not coalesce the Spark DataFrame in one partition before writing the text file. This is the only option when the resulting DataFrame is too large to save it in one local partition but is not so convenient for future processing:

%%time
df_sample.write.csv('sample.csv') # Needs Spark >= 2.0. In the cluster, we have Spark 1.5.0

%%time
df_sample.write.format('com.databricks.spark.csv').save('sample2.csv') # For Spark 1.5.0

In order to save the data in one file, we can use these handier alternatives. Our first aggregations of the full dataset will be small enough (~ 80k lines, ~500k lines) to be able to do this:

%%time
df_sample.rdd.map(lambda x: ','.join(map(str, x))).coalesce(1).saveAsTextFile('file_rdd.csv')

%%time
df_sample.toPandas().to_csv('file_pd.csv')

Transforming it to a Pandas DataFrame was quicker, simpler and the exit file includes the header. We will follow this path when data is small enough.

#### Processing the full dataset

Now let us start tackling the real thing:

path = 'Power-Networks-LCL-June2015.bz2'

data = sc.textFile(path)

data.take(5)

There are two new problems.

First of all, there are two extra lines above the header, so now we need to remove 3 lines before mapping.

data_no_header = data.zipWithIndex().filter(lambda x: x[1] > 2).keys()

data_no_header.map(line2tuple).take(3)

Moreover, the DateTime format has changed compared to the sample provided!! Not very nice, let's hope that at least it is consistent thoughout the whole full dataset.
We must change our mapping function accordingly. This datetime format can be automatically changed later for the full dataset to pyspark.sql TimestampType so we will leave it as string for now.

data_no_header.map(line2tuple).filter(lambda (_1,t,_2,_3,_4): t == 'ToU').take(3)

data_no_header.map(line2tuple)\
.filter(lambda (u,t,_2,_3,_4): (u == 'MAC000005') & (t=='ToU')).take(3)

So actually "stdorToU" tells us if a user was subjected to dToU or std tariff in 2013, not in the timestamp corresponding to a given measurement. We do not need to separate users then but we should bear in mind later that dToU tariff was only applied in 2013.

def line2tuple(l):
    fields = l.split(',')
    if len(fields) == 6: 

#We have seen that there is at least one line in the file that does not follow the format.
#This condition is to avoid an "Index out of range" error. 
        ID = fields[0]
        tariff = fields[1]
        DateTime = fields[2]
        ACORN = fields[4]
        try:
            consumption = float(fields[3])
        except ValueError:
            consumption = np.nan
        return (ID, tariff, DateTime, consumption,ACORN)
    else:
        return (np.nan,np.nan,np.nan,np.nan,np.nan)
   
#We treat any line that does not follow the general format as NaN. 
#We will remove them afterwards

We create our dataframe and check that DateTime is a string variable.

rows = data_no_header.map(line2tuple)\
.map(lambda x: Row(ID = x[0], Tariff = x[1], DateTime = x[2], kWh_30min = x[3]))

df = rows.toDF()

df.printSchema

We change DateTime to pyspark.sql TimestampType (this is equivalent to datetime.datetime).

from pyspark.sql.types import TimestampType
df2 = df.withColumn('DateTime', df['DateTime'].cast(TimestampType()))

df2.printSchema

df2.show(5)

Let's try if calculating the mean grouped by tariff works, this would mean we have cleaned the dataset successfully.

df2.dropna().groupBy('Tariff').mean().collect()

That's great! We have also learnt the keys used to name the two tariff groups we expected to find ('Std' and 'ToU')

It will be useful in our analysis which consumers were subjected to Dynamic Time of Use (ToU) tariff during 2013. These had a standard tariff during the rest of the period included in the dataset so we can evaluate behavioural changes due to the tariff scheme.

df2.filter(df2['Tariff'] == 'ToU').select('ID').drop_duplicates()

In the first step of our analysis we will aggregate all users under the same tariff plan (distinguishing consumers subjected to ToU during 2013 and consumers with the standard flat rate tariff throughout the whole period of time). We will analyse this data in R later (no need for Spark), including explanatory plots. R provides useful packages for this purpose: xts and zoo for time series plotting and forecast for prediction models.

Therefore, we will calculate basic statistics for each timestamp per tariff plan. We will then refine the analysis aggregating by ACORN group and tariff plan.

def ACORN(l):
    fields = l.split(',')
    if len(fields) == 6:
        acorn = fields[4]
        return acorn
    else:
        return 'NA'

#ACORN groups
ACORN_groups = data_no_header.map(ACORN).filter(lambda a: a != 'NA').distinct()

ACORN_groups.collect()

Code for script to be executed in the cluster:

#from pyspark import SparkContext
#sc = SparkContext()
from pyspark.sql.types import DateType
from datetime import datetime
import numpy as np
from pyspark import Row

path = 'Power-Networks-LCL-June2015.bz2'

data = sc.textFile(path)
data_no_header = data.zipWithIndex().filter(lambda x: x[1] > 2).keys()

def ToU_IDs(l):
    fields = l.split(',')
    if len(fields) == 6:
        ID = fields[0]
        tariff = fields[1]
        return (tariff, ID)
    else:
        return ('NA','NA')

ID_ToU_rdd = data_no_header.map(ToU_IDs).filter(lambda (t,_): t == 'ToU').distinct()
ID_ToU = ID_ToU_rdd.map(lambda (_,i): i).collect()

def line2tuple(l):
    fields = l.split(',')
    if len(fields) == 6:
        ID = fields[0]
        tariff = fields[1]
        DateTime = fields[2]
        ACORN = fields[4]
        try:
            consumption = float(fields[3])
        except ValueError:
            consumption = np.nan
        return (ID, tariff, DateTime, consumption, ACORN)
    else:
        return (np.nan,np.nan,np.nan,np.nan,np.nan)

rows = data_no_header.map(line2tuple)\
.map(lambda x: Row(ID = x[0], Tariff = x[1], DateTime = x[2], 
                   kWh_30min = x[3], ACORN = x[4]))

df = rows.toDF()

df.show(5)

from pyspark.sql.types import TimestampType
df = df.withColumn('DateTime', df['DateTime'].cast(TimestampType()))

df.printSchema

import pyspark.sql.functions as F

agg_stats = df.dropna().groupBy('DateTime','Tariff')\
    .agg(F.count('kWh_30min').alias('count'),
         F.sum('kWh_30min').alias('sum'),
         F.min('kWh_30min').alias('min'),
         F.mean('kWh_30min').alias('mean'),
         F.max('kWh_30min').alias('max'),
         F.stddev('kWh_30min').alias('std_dev')
        )

As the grouped Dataframe is small enough to be saved in one partition we can transform it to a Pandas dataframe and then save it to csv. Saving it in only one text file is much handier for future processing.

We could have also transformed it back to a functional API rdd and used coalesce(1).saveAsTextFile('agg_stats.csv').

Otherwise, we could have written it to multiple text files with:

agg_stats.write.csv('agg_stats.csv') (Spark >= 2.0)

agg_stats.write.format('com.databricks.spark.csv').save('agg_stats.csv') (in the cluster, where we have Spark 1.5.0)

agg_stats.toPandas().to_csv('outputs/agg_stats.csv')

acorn_stats = df.dropna().groupBy('ACORN','DateTime','Tariff')\
    .agg(F.count('kWh_30min').alias('count'),
         F.sum('kWh_30min').alias('sum'),
         F.min('kWh_30min').alias('min'),
         F.mean('kWh_30min').alias('mean'),
         F.max('kWh_30min').alias('max'),
         F.stddev('kWh_30min').alias('std_dev')
        )

acorn_stats.toPandas().to_csv('outputs/acorn_stats.csv')

We will copy this last part to SmartMeter_agg1.py to run it in the cluster.
