==> 4.1_Aggregate_demand_prediction-DS_Preparation.ipynb <==
# 4 Prediction of the aggregate demand 

In this section, we will compare the performance of different algorithms in predicting the aggregate demand, separating standard and ToU users. First, we will use feature models and then exponential smoothing, which is a time-series dedicated algorithm.

After adding the external variables (holidays, weather) we will use different algorithms to fit our feature model. Later, we will predict the aggregate demand with exponential smoothing models in R using the "forecast" package.

In general we will train our model with Nov 2011- Nov 2013 data and validate it with Nov 2013 - Feb 2014 data.

We will use the aggregate statistics we computed in Section 2 with Spark.

## 4.1 Preparing the dataset

First, we will prepare our data so it can be use in the different algorithms proposed. We will start including time and weather variables as these are needed for all models and then we will add the engineered features for the feature models.

import pandas as pd
pd.options.display.max_columns = 999
import datetime

df = pd.read_csv('outputs/agg_stats.csv', index_col=0)

df.head()

We keep only the columns we need: DateTime, Tariff and mean consumption. We are interested in predicting the aggregate demand (i.e. the sum) but as the data count in each Timestamp is different, we will use the mean. However, we cannot take the mean column directly as the value of Tariff is the user group rather than the tariff applied. In 2011, 2012 and 2014 we are interested in the aggregate demand of both groups combined so we will group and calculate the mean with sum and count.

df = df[['DateTime','Tariff','sum','count']]

df['DateTime'] = pd.to_datetime(df['DateTime'])

type(df['DateTime'][2])

Column Tariff actually represents the user group, not the tariff applied in that timestamp so we will rename it for clarity.

df = df.rename(columns={'Tariff':'User_group'})

df.sample(5)

### 4.1.1 Tariff

On of the main variables to include in our dataset is the price of electricity. We have grouped standard flat rate and dToU users in two different groups but we need the actual values of the tariff. 

For standard users, the price is always equal to 14.228 p/kWh.

For ToU users, the tariff level for each 30-min interval is given in an Excel sheet we downloaded in Section 3.2. The corresponding values for each level are:

-High = 67.20p/kWh

-Normal = 11.76p/kWh

-Low = 3.99p/kWh

tariffs = pd.read_excel('data/Tariffs.xlsx')

tariffs.loc[tariffs['Tariff'] == 'Normal', 'Tariff_value'] = 11.76
tariffs.loc[tariffs['Tariff'] == 'High', 'Tariff_value'] = 67.20
tariffs.loc[tariffs['Tariff'] == 'Low', 'Tariff_value'] = 3.99

tariffs = tariffs.rename(columns = {'TariffDateTime' : 'DateTime'})

tariffs.sample(5)

tariffs = tariffs.drop('Tariff', axis=1)

df = df.merge(tariffs, how = 'left', on = 'DateTime')

df['Tariff_value'] = df['Tariff_value'].fillna(14.228) #This is the standard tariff

df['Tariff_value'] = df['Tariff_value'].fillna(14.228) #This is the standard tariff
df.loc[df['User_group'] == 'Std','Tariff_value'] = 14.228

df.sample(5)

df = df.groupby(['DateTime', 'Tariff_value']).agg({'sum': 'sum',
                                            'count': 'sum'}).reset_index()

Now we can calculate the mean for each data point.

df['mean'] = df['sum']/df['count']

We also add a new variable "ToU" which indicates whether a consumption data point is for a user in which ToU was applied. As 14.228 p/kWh is the flat rate tariff (and this is not a possible value of ToU tariff), we know that if Tariff_value = 14.228, this is a standard point, and otherwise a ToU point.

df.loc[df['Tariff_value'] == 14.228,'ToU'] = 0
df.loc[df['Tariff_value'] != 14.228,'ToU'] = 1

df.sample(5)

### 4.1.2. Bank Holidays

Bank Holidays probably have an impact in household electricity consumption as it was discussed in Section 1. Bank Holidays will be treated separately from Sundays as a binary variable. 

Bank Holidays in England from 2011 to 2014 were downloaded in Section 1 using the Python holidays library (and cross-checked with the official information in www.gov.co.uk) and saved to a csv file.

holidays = pd.Series.from_csv('data/bank_holidays.csv')

holidays = pd.to_datetime(holidays)

holidays.head()

df['Date'] = df['DateTime'].apply(lambda x: x.date())

df['Date'] = pd.to_datetime(df['Date'])

df['Holiday'] = 0
df.loc[df['Date'].isin(holidays),'Holiday'] = 1

df.sample(5)

df[df['Holiday'] == 1].sample(5)

### 4.1.3. Weather variables

Now we have to add the weather variables, which will be used in any model. Weather variables were discussed, downloaded and treated in Section 1.

weather = pd.read_csv('data/weather_no_na.csv')

weather.shape

We will use the following variables:

-Temperature (ºC)

-Relative Humidity (%)

-Cloud cover (%)

-Atmospheric Pressure (mbar)

For a discussion on the reasons why these variables have been choosen and how NA values have been filled refer to Section 1.

weather.columns

weather = weather[['time','temperature','humidity','cloudCover','pressure']]\
    .rename(columns = {'time' : 'DateTime'})

weather['DateTime'] = pd.to_datetime(weather['DateTime'])

weather.head()

The frequency of the weather data is of 1 hour while that of the consumption data is 30 minutes, thus we need to resample it. We will take the last valid observation (e.g. 2012-01-01 10:30 will have the same data as 2012-01-01 10:00) => method = 'ffill'.

weather = weather.set_index('DateTime').resample('30min').fillna(method = 'ffill')\
    .reset_index()

weather.head()

df = df.merge(weather, how = 'left', on = 'DateTime')

df.sample(5)

Now that we have added all external variables our models need, we can save it to a csv file. 

We will split the dataset into train and test samples: from the dataset start to 2013-09-31 will be included in the train set and from 2013-10-01 until the end of the time series in the test set. This means we will have 9 months of ToU in train and 3 in test. The limitation is not all months will be included in the test set and we will not have one whole year of ToU data in the train set.

df.loc[df['DateTime'] < '2013-10-01 00:00:00','Set'] = 'Train'
df.loc[df['DateTime'] >= '2013-10-01 00:00:00','Set'] = 'Test'

df.to_csv('outputs/tseries_model.csv', index = False)

We will use this model as is in time series modelling. For feature models we need to add the features.

## 4.1.4. Feature engineering

Feature models are an alternative for time series forecasting. They consist in applying conventional machine learning models to variables constructed from the data (e.g. mean consumption in the previous day). These are popular in the literature for electrical demand prediction [2], [4].

We will derive the following features:

-Mean aggregate demand per hour in the previous day (1 variable).

-Aggregate demand in the previous day at the same time and at the previous 3 time steps (4 variables).

-Mean value of the aggregate demand of the previous 3 days at the same time and at the previous 3 time steps (4 variables).

-Mean value of the aggregate demand on the same day of week of the previous 3 weeks at the same time and at the previous 3 time steps (4 variables).

In total we have 13 derived features.

Furthermore, we will include weather and time variables in the model:

-Day of week (7 levels)

-Holiday (binary)

-Temperature (ºC, continuous)

-Relative Humidity (%, continuous)

-Cloud cover (%, continuous)

-Atmospheric Pressure (mbar, continuous)

These variables have been calculated in Section 1 and stored in 'weather_no_na.csv'.

And the last variable to be included in the model:

-Tariff (p/kWh)

Thus, in total we will have 22 variables in our models.


First, we will build the time variables.

Now it is easier to derive the consumption features.

#### Adding time variables

Let us add the time variables to be used in lieu of DateTime in our prediction model: time, day of week, month, year.

We will not use the day of the month as it would introduce a categorical variable with 31 levels and it does not seem relevant in the exploratory data analysis in Tableau.

df['DoW'] = df['DateTime'].apply(lambda x: x.strftime("%A"))

df['Time'] = df['DateTime'].apply(lambda x: x.time())

df['Month']= df['DateTime'].apply(lambda x: x.strftime("%B"))

df['Year']= df['DateTime'].apply(lambda x: x.year)

df.sample(5)

###### Mean aggregate demand per hour in the previous day (1 variable).

First, we will calculate the daily mean aggregating by 'Date'. Then we will add one day to the 'Date' variable of our new DataFrame and join it with the main DataFrame df.

daily_mean = df.groupby(['Date', 'ToU']).agg({'mean': 'mean'}).reset_index()

daily_mean['Date'] = pd.to_datetime(daily_mean['Date'])

type(daily_mean['Date'][2])

daily_mean.sample(5)

daily_mean['Date'] = daily_mean['Date'] + datetime.timedelta(days=1)

df = df.merge(daily_mean, how = 'left', on = ['Date','ToU'])

df = df.rename(columns = {'mean_x' : 'mean_cons',
          'mean_y' : 'mean_prev_day'})

df.sample(5)

###### Aggregate demand in the previous day at the same time and at the previous 3 time steps (4 variables).

df.head()

def prev_ts(df_in, initial, final, step):
    #Both initial and final are included in the loop
    df_out = df_in[['DateTime', 'ToU']].copy()
    for i in range(initial, final + 1, step):
        aux = df_in[['DateTime', 'ToU', 'mean_cons']].copy()
        aux['DateTime'] = aux['DateTime'] + datetime.timedelta(minutes = i)
        df_out = df_out.\
            merge(aux, how = 'left', on = ['DateTime','ToU'], suffixes = ('','_-%d' %(i)))
    return(df_out)

tmp = prev_ts(df, 22*60+30, 24*60, 30)
df = df.merge(tmp[['DateTime','ToU','mean_cons_-1380', 'mean_cons_-1410', 'mean_cons_-1440']],
         on = ['DateTime','ToU'], suffixes = ('',''))

df.sample(5)

###### Mean value of the aggregate demand of the previous 3 days at the same time and at the previous 3 time steps (4 variables).

aux2 = df[['DateTime', 'ToU']].copy()
means_3d = pd.DataFrame(index=df.index)
for h in range(0,4):
    for d in range(1,4):
        aux = df[['DateTime', 'ToU', 'mean_cons']].copy()
        aux['DateTime'] = aux['DateTime'] + datetime.timedelta(minutes = d*24*60-h*30)
        aux2 = aux2.\
            merge(aux, how = 'left', on = ['DateTime','ToU'])
    means_3d = means_3d.join(aux2.mean(axis=1).rename('mean_last3d_-%d' %(h*30)))

df = df.join(means_3d)

df.sample(3)

###### Mean value of the aggregate demand on the same day of week of the previous 3 weeks at the same time and at the previous 3 time steps (4 variables).

aux2 = df[['DateTime', 'ToU']].copy()
means_3w = pd.DataFrame(index=df.index)
for h in range(0,4):
    for w in range(1,4):
        aux = df[['DateTime', 'ToU', 'mean_cons']].copy()
        aux['DateTime'] = aux['DateTime'] + datetime.timedelta(minutes = w*7*24*60-h*30)
        aux2 = aux2.\
            merge(aux, how = 'left', on = ['DateTime','ToU'])
    means_3w = means_3w.join(aux2.mean(axis=1).rename('mean_last3w_-%d' %(h*30)))

df = df.join(means_3w)

df.sample(3)

Note that pandas.dataframe.mean skips NaN values by default. Therefore, if some values are missing, the mean is still calculated with the remaining ones. Nonetheless, in the first 3 weeks, none of the values needed to calculate some features will exist so there will be missing values. These values will be dropped from the model. This is not needed in dedicated time series algorithms.

df = df.dropna().reset_index(drop=True)

Time can be encoded as a categorical variable or as a continuous numerical variable.

Looking at the exploratory data analysis in Tableau, aggregating data by time, we can distinguish a plateau between 8:00 and 15:59 during weekdays, an increasing trend up to 19:00, then decreasing to 2:00 and a plateau between 2:00 and 5:59 before increasing to 8:00. During weekends, the midday plateau is not so constant and seems to start at 9:00.

Considering time with 30-min frequency would result in a categorical variable with 48 levels. This resolution is probably to large for our purpose taking into account that we will derive features based on consumption in previous time steps and at the same time on previous comparable days, as we will see later. 

Considering time as a continuous numerical variable is probably a better approach, but from the exploratory data analysis we now its relationship with the target will be strongly non-linear.

Therefore, we propose to use it as a numerical variable in the interval [0,24) in models which can grasp non-linearities and to drop it in linear models at is has been used to derived the features.

We will also remove the year as we only have 2 complete years (2012 and 2013) and part of our test data is in 2014, which is absent from the train set.

def time2float(time):
    return time.hour + time.minute / 60.0

df['Time'] = df['Time'].apply(lambda t: time2float(t))

Now that we have built all features we can save this model to a csv file for use in R.

df.to_csv('outputs/features_model.csv', index = False)
